uint -> uint256 (#460)

+++ contracts/margin/external/exchangewrappers/ZeroExV1ExchangeWrapper.sol
@@ -85,7 +85,7 @@ contract ZeroExV1ExchangeWrapper is
         ZERO_EX_TOKEN_PROXY = zeroExProxy;
         ZRX = zrxToken;
 
-        for (uint i = 0; i < trustedMsgSenders.length; i++) {
+        for (uint256 i = 0; i < trustedMsgSenders.length; i++) {
             TRUSTED_MSG_SENDER[trustedMsgSenders[i]] = true;
         }
